+++ src/datasources/results/components/variable-editors/ResultsVariableQueryEditor.test.tsx
@@ -1,10 +1,11 @@
-import { screen, waitFor } from '@testing-library/react';
+import { fireEvent, screen, waitFor } from '@testing-library/react';
 import { ResultsVariableQueryEditor } from './ResultsVariableQueryEditor';
 import { setupRenderer } from 'test/fixtures';
 import { ResultsDataSource } from 'datasources/results/ResultsDataSource';
 import { ResultsQuery } from 'datasources/results/types/types';
 import { Workspace } from 'core/types';
 import { QueryResultsDataSource } from 'datasources/results/query-handlers/query-results/QueryResultsDataSource';
+import { ResultsVariableProperties } from 'datasources/results/types/QueryResults.types';
 
 const fakeWorkspaces: Workspace[] = [
   {
@@ -40,12 +41,18 @@ const renderEditor = setupRenderer(ResultsVariableQueryEditor, FakeResultsDataSo
 let propertiesSelect: HTMLElement;
 let queryBy: HTMLElement;
 
-it('should render properties select and results query builder', async () => {
+it('should render properties select and results query builder progressively', async () => {
   renderEditor({ refId: '', properties: '', queryBy: '' } as unknown as ResultsQuery);
   propertiesSelect = screen.getAllByRole('combobox')[0];
-  queryBy = screen.getByText('Query By');
 
   expect(propertiesSelect).toBeInTheDocument();
+
+  //simulate user selecting a property
+  fireEvent.keyDown(propertiesSelect, { key: 'ArrowDown' });
+  const option = await screen.findByText(ResultsVariableProperties[0].label);
+  fireEvent.click(option);
+
+  queryBy = screen.getByText('Query by results properties');
   expect(queryBy).toBeInTheDocument();
 
   await waitFor(() => expect(screen.getAllByText('Property').length).toBe(1));